update install instructions

+++ install.sh
@@ -15,22 +15,27 @@ shell_join() {
   done
 }
 
+abort() {
+  printf "%s\n" "$1"
+  exit 1
+}
+
 execute() {
   if ! "$@"; then
     abort "$(printf "Failed during: %s" "$(shell_join "$@")")"
   fi
 }
 
+if [ -d "${SPLINTER_DIR}" ];then
+  abort "$(printf "Failed directory %s already exists!" "$(shell_join "${SPLINTER_DIR}")")"
+fi
+
 echo "Downloading Splinter in to '${SPLINTER_DIR}'..."
-(
-  # we do it in four steps to avoid merge errors when reinstalling
-  execute "git" "clone" "-q" "${SPLINTER_REPO}"
-
-  # remove .git dir
-  cd "${SPLINTER_DIR}" || exit
-  rm -rf '.git' || exit
-)
-
+# we do it in four steps to avoid merge errors when reinstalling
+execute "git" "clone" "-q" "${SPLINTER_REPO}" || exit 1
+# remove .git dir
+cd "${SPLINTER_DIR}" || exit
+rm -rf '.git' || exit
 echo "Installation successful!"
 
 if [[ "$(uname)" = "Darwin" ]]; then
